字体颜色和字体逻辑已经准备好啦，现在还是对其时间线

+++ CutSameUIIF/src/main/java/com/cutsame/ui/cut/lyrics/colorselect/ColorCircleView.kt
@@ -39,8 +39,8 @@ class ColorCircleView : View {
         canvas.drawCircle(radius.toFloat(), radius.toFloat(), radius.toFloat(), fillPaint)
     }
 
-    fun setColor(@ColorInt color: Int) {
-        fillColor = color
+    fun setColor(@ColorInt color: Long) {
+        fillColor = color.toInt()
         fillPaint.color = fillColor
     }
